+++ lib/src/string.dart
@@ -5,6 +5,14 @@ import 'json.dart';
 const _encoder = JsonEncoder.withIndent('  ');
 
 /// Calls [dumpJson] and pretty-prints the result to a string.
-String dumpString(Object? value) {
-  return _encoder.convert(dumpJson(value));
+String dumpString(
+  Object? value, {
+  Set<String> hideFields = const {},
+}) {
+  return _encoder.convert(
+    dumpJson(
+      value,
+      hideFields: hideFields,
+    ),
+  );
 }